authorizated 미들웨어 수정

+++ services/cards.service.js
@@ -1,12 +1,19 @@
 const CardRepo = require('../repositories/cards.repository');
 const { Transaction } = require('sequelize');
-const { MakeError } = require('../utils/makeErrorUtil');
+const MakeError = require('../utils/makeErrorUtil');
 
 class CardService {
   cardRepo = new CardRepo();
 
-  createCard = async (columnId, description, deadline) => {
+  createCard = async (columnId, description, deadline, manager) => {
     try {
+      if (!columnId) {
+        throw new MakeError(400, 'columnId가 존재하지 않습니다.');
+      }
+      if (!description) {
+        throw new MakeError(400, 'description은 필수 입력사항입니다.');
+      }
+
       const createCardResult = await this.cardRepo.createCard(
         columnId,
         description,
@@ -24,6 +31,9 @@ class CardService {
 
   getAllCards = async (columnId) => {
     try {
+      if (!columnId) {
+        throw new MakeError(400, 'columnId가 존재하지 않습니다.');
+      }
       const getCardsResult = await this.cardRepo.getAllCards(columnId);
 
       if (!getCardsResult) {
@@ -46,6 +56,10 @@ class CardService {
     manager,
   ) => {
     try {
+      if (!cardId) {
+        throw new MakeError(400, 'cardId를 수신받지 못했습니다.');
+      }
+
       const updateCardResult = await this.cardRepo.modifyCard(
         cardId,
         name,
@@ -71,6 +85,9 @@ class CardService {
       isolationLevel: Transaction.ISOLATION_LEVELS.READ_COMMITTED,
     });
     try {
+      if (!positionInfos.length === 0) {
+        throw new MakeError(400, 'positionInfos가 존재하지 않습니다.');
+      }
       const resultFirst = await this.cardRepo.modifyCardPosition(
         positionInfos[0],
         t,
@@ -94,6 +111,9 @@ class CardService {
 
   deleteCard = async (cardId) => {
     try {
+      if (!cardId) {
+        throw new MakeError(400, 'cardId를 수신받지 못했습니다.');
+      }
       const deleteCardResult = await this.cardRepo.deleteCard(cardId);
       if (!deleteCardResult) {
         throw new MakeError('402', '해당 카드를 삭제하지 못했습니다.');
